requirements.txt
pytorch==2.0.1
monai==0.8.0
simpleitk==2.1.1
nibabel==3.2.1
natsort==7.0.1
plotly==5.5.0
numpy==1.23.5
pandas==1.5.3
natsort==8.3.1
matplotlib==3.7.0